Update @github/copilot to 1.0.51-1 (#1340)
* Update @github/copilot to 1.0.51-1

- Updated nodejs and test harness dependencies
- Re-ran code generators
- Formatted generated code

* Fix SDK codegen for Copilot 1.0.51-1


* Address SDK review feedback


* Fix SDK compatibility with CLI 1.0.51


* Fix post-tool hook snapshot parity


* Fix Rust hook e2e test isolation


---------

+++ go/internal/e2e/rpc_event_side_effects_e2e_test.go
@@ -199,7 +199,7 @@ func TestRpcEventSideEffectsE2E(t *testing.T) {
 		if !strings.EqualFold(rewindData.UpToEventID, targetEventID) {
 			t.Fatalf("Expected rewind to target %q, got %+v", targetEventID, rewindData)
 		}
-		if rewindData.EventsRemoved != float64(truncateResult.EventsRemoved) {
+		if rewindData.EventsRemoved != truncateResult.EventsRemoved {
 			t.Fatalf("Expected rewind count %d, got %+v", truncateResult.EventsRemoved, rewindData)
 		}
